add hamiltonian script

+++ all_hamiltonian_or_simplified_sat.sh
@@ -2,8 +2,9 @@
 
 # Load FF table
 source ./ff_table.sh
+source ./pi_table.sh
 
-dir_path="iscas89/result/hamiltonian_simplified/"
+dir_path="iscas89/result/hamiltonian_or_simplified/"
 
 mkdir -p "$dir_path"
 rm iscas89/benchmarks/*simplified.*
@@ -42,12 +43,21 @@ for (( FF_num=3; FF_num<=3; FF_num=FF_num+1 )); do
 			echo "$ins skip"
 			continue
 		fi
+		pi_threshold=${pi_table[$ins]}
+		if [ -z "$pi_threshold" ]; then
+			echo "No PI number found for $ins, skipping..."
+			continue
+		fi
+		if (( pi_threshold+FF_num > 30 )); then
+			echo "$ins skip"
+			continue
+		fi
 
 		echo $bench &> "$file_name"
 		echo "${ins}" &>> "$file_name"
 		echo "FF_tokeep=${FF_num}" &>> "$file_name"
 
-		./my_hamiltonian_simplified_iscas.sh "$ins" "$FF_num" &>> "$file_name"
+		./or_hamiltonian_simplified_iscas.sh "$ins" "$FF_num" &>> "$file_name"
 
 		rm iscas89/benchmarks/*simplified.*
 	done
